Make cleanup scripts skip verification_tokens when table is missing.

+++ scripts/cleanup-all-bookings.sql
@@ -2,7 +2,14 @@ BEGIN;
 
 DELETE FROM "web_bookings";
 DELETE FROM "app_alerts";
-DELETE FROM "verification_tokens";
+
+DO $$
+BEGIN
+  IF to_regclass('public.verification_tokens') IS NOT NULL THEN
+    DELETE FROM "verification_tokens";
+  END IF;
+END $$;
+
 DELETE FROM "notification_outbox";
 
 COMMIT;
